requirements.txt
# requirements.txt
fastapi
uvicorn
anthropic
mcp
python-dotenv        # ✅ Add this